+++ src/core/index.ts
@@ -14,6 +14,7 @@ export * from "./rank";
 export * from "./ids";
 export * from "./lookup";
 export * from "./query";
+export * from "./resource";
 export * from "./errors";
 export * from "./status";
 export * from "./intent";